+++ RELEASE_NOTES.md
@@ -149,9 +149,8 @@ scripts.
 
 * ADDED: Sample Bash shell embedding script for YAP.
 
-* UPDATED: The sample embedding scripts for ECLiPSe, GNU Prolog, JIProlog,
-SICStus Prolog, and SWI-Prolog to also allow embedding an optional settings
-file and an optional application.
+* UPDATED: The sample embedding scripts to also allow embedding an optional
+settings file and an optional application.
 
 * UPDATED: The sample embedding scripts to allow selecting a library paths
 file other than the `paths_core.pl` default file.